+++ bevy_ninepatch/examples/multi_content.rs
@@ -2,7 +2,6 @@ use bevy::prelude::*;
 
 use bevy_ninepatch::{
     NinePatchBuilder, NinePatchComponents, NinePatchContent, NinePatchData, NinePatchPlugin,
-    NinePatchSize,
 };
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -43,6 +42,7 @@ fn setup(
                 margin: Rect::all(Val::Auto),
                 justify_content: JustifyContent::Center,
                 align_items: AlignItems::Center,
+                size: Size::new(Val::Px(500.), Val::Px(300.)),
                 ..Default::default()
             },
             nine_patch_data: NinePatchData {
@@ -50,7 +50,6 @@ fn setup(
                 texture: panel_texture_handle,
                 ..Default::default()
             },
-            nine_patch_size: NinePatchSize(Vec2::new(500., 300.)),
             ..Default::default()
         },
     );
@@ -95,6 +94,7 @@ fn set_content(
                             },
                             justify_content: JustifyContent::Center,
                             align_items: AlignItems::Center,
+                            size: Size::new(Val::Px(200.), Val::Px(100.)),
                             ..Default::default()
                         },
                         nine_patch_data: NinePatchData {
@@ -102,7 +102,6 @@ fn set_content(
                             texture: button_texture_handle,
                             ..Default::default()
                         },
-                        nine_patch_size: NinePatchSize(Vec2::new(200., 100.)),
                         ..Default::default()
                     },
                 );