+++ bevy_ninepatch/examples/plugin.rs
@@ -1,8 +1,6 @@
 use bevy::prelude::*;
 
-use bevy_ninepatch::{
-    NinePatchBuilder, NinePatchComponents, NinePatchData, NinePatchPlugin, NinePatchSize,
-};
+use bevy_ninepatch::{NinePatchBuilder, NinePatchComponents, NinePatchData, NinePatchPlugin};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     App::build()
@@ -36,6 +34,7 @@ fn setup(
                 margin: Rect::all(Val::Auto),
                 justify_content: JustifyContent::Center,
                 align_items: AlignItems::Center,
+                size: Size::new(Val::Px(500.), Val::Px(300.)),
                 ..Default::default()
             },
             nine_patch_data: NinePatchData {
@@ -43,7 +42,6 @@ fn setup(
                 texture: panel_texture_handle,
                 ..Default::default()
             },
-            nine_patch_size: NinePatchSize(Vec2::new(500., 300.)),
             ..Default::default()
         },
     );